+++ packages/jupyterlab-rjsf/src/form/Form/Form.tsx
@@ -6,11 +6,8 @@ import { asyncComponent } from '../../async-component';
  * The actual form/theme will be loaded asynchronously the first time it is used
  */
 const Form = asyncComponent(async () => {
-  const { withTheme } = await import(
-    /* webpackChunkName: "@rjsf/core" */ '@rjsf/core'
-  );
-  const Theme = (await import(/* webpackChunkName: "@rjsf/core" */ '../Theme'))
-    .default;
+  const { withTheme } = await import('@rjsf/core');
+  const Theme = (await import('../Theme')).default;
   return withTheme(Theme);
 });